Changed Lombok annotation from Builder to SuperBuilder for abstract classes and their inheritors, added tests for combustion engine repository

+++ src/main/java/u238/reauto/repository/vehicle/parts/engine/CombustionEngineRepository.java
@@ -11,6 +11,6 @@ public interface CombustionEngineRepository extends DefaultEngineRepository<Comb
 
     List<CombustionEngine> findAllByCylinders(int cylinderNumber);
 
-    List<CombustionEngine> findAllBySizeLiters(int sizeLiters);
+    List<CombustionEngine> findAllBySizeLiters(double sizeLiters);
 
 }
